requirements.txt
streamlit==1.28.0
python-docx==0.8.11
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
pillow==10.0.1
sqlalchemy==2.0.23
python-magic==0.4.27
spacy==3.7.2
python-multipart==0.0.6
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
coverage==7.3.2